Move duplicated ColorPaletteRepositoryInterface out of the two palette use cases and into the use_case folder
